organized component file structure
