+++ src/components/Companies/View/CompanyDetail.js
