+++ src/components/Shift/Positions/PositionDetail.js
