+++ src/components/Shift/View/Map.js
