+++ src/components/Shift/View/ShiftDetail.js
