+++ src/components/Workforce/View/WorkerDetail.js
